Fix issues with Bundle environment leakage.
NOTE: The old version of this code was broken, and I thought the fix would just
be to pass an empty array to Bundler::CLI.start, but I uncovered more issues
when I did this, so this jumps straight to a bit more convoluted fix.

Previously, our own implied use of Bundler would leak into the
Bundler::CLI.start usage, and cause wonky behavior. Bundler ignored our
APP_CONFIG, because that option is only read the first time anyone calls
Bundler.settings, which Bundler does internally.

Instead of doing this, we now call out with system() to bundler, but this means
that our generate_standalone method no longer works. Because of this, we now
depend on bundler 1.1.rc, and use the --standalone option.

Depending on 1.1.rc means that to develop with it you need to manually gem
install --pre bundler (I couldn't find a way to get the Gemfile to install it).
People using the gem will automatically have bundler 1.1.rc installed, though.

Finally, this fixes the cache ignore code a little bit.