Fix loading of Ruby plugins.
In Jenkins 1.519 a change was made to move plugin classes into
WEB-INF/lib/classes.jar rather than unpacked into WEB-INF/classes/.

This change causes the ruby runtime to unpack the classes.jar file to
WEB-INF/lib/classes/ when it loads a ruby plugin. It also changes the
load paths to look for the extra /lib in the file path.

Fixes JENKINS-18528.